requirements.txt
torch >= 2.0.0
transformers >= 4.40.0
accelerate >= 0.30.0
huggingface-hub >= 0.23.0
sentence-transformers >= 2.7.0
langchain >= 0.2.0
langchain-core >= 0.2.0
langchain-community >= 0.1.0
langchain-text-splitters >= 0.2.0
chromadb >= 0.5.0
langchain-chroma >= 0.2.0
langserve[all] >= 0.1.0
fastapi >= 0.115.0
uvicorn >= 0.30.0
gradio >= 5.0.0
langchain-huggingface
streamlit
faiss-cpu
numpy
groq